kdf-components/src/NIST.CVP.Generation.SNMP/TestGroup.cs
using System.Collections.Generic;
using System.Dynamic;
using System.Linq;
using Newtonsoft.Json.Linq;
using NIST.CVP.Generation.Core;
using NIST.CVP.Generation.Core.ExtensionMethods;
using NIST.CVP.Math;

namespace NIST.CVP.Generation.SNMP
{
    public class TestGroup : ITestGroup
    {
        public TestGroup()
        {
            Tests = new List<ITestCase>();
        }

        public TestGroup(JObject source) : this(source.ToObject<ExpandoObject>()) { }

        public TestGroup(dynamic source)
        {
            var expandoSource = (ExpandoObject) source;

            TestGroupId = expandoSource.GetTypeFromProperty<int>("tgId");
            EngineId = expandoSource.GetBitStringFromProperty("engineId");
            PasswordLength = expandoSource.GetTypeFromProperty<int>("passwordLength");

            Tests = new List<ITestCase>();
            foreach (var test in source.tests)
            {
                Tests.Add(new TestCase(test));
            }
        }

        public int TestGroupId { get; set; }
        public BitString EngineId { get; set; }
        public int PasswordLength { get; set; }

        public string TestType { get; set; }
        public List<ITestCase> Tests { get; set; }

        public bool SetString(string name, string value)
        {
            if (string.IsNullOrEmpty(name))
            {
                return false;
            }

            switch (name.ToLower())
            {
                case "engineid":
                    EngineId = new BitString(value);
                    return true;
                
                case "passwordlen":
                    PasswordLength = int.Parse(value);
                    return true;
            }

            return false;
        }
    }
}
